moved all avalon context functions to one file
