+++ pype/lib/__init__.py
@@ -6,6 +6,17 @@ from .deprecated import (
     set_io_database
 )
 
+from .avalon_context import (
+    is_latest,
+    any_outdated,
+    switch_item,
+    get_asset,
+    get_hierarchy,
+    get_subsets,
+    get_linked_assets,
+    get_latest_version
+)
+
 from .hooks import PypeHook, execute_hook
 
 from .applications import (
@@ -20,31 +31,32 @@ from .lib_old import (
     _subprocess,
     get_paths_from_environ,
     get_ffmpeg_tool_path,
-    get_hierarchy,
     add_tool_to_environment,
     modified_environ,
     pairwise,
     grouper,
-    is_latest,
-    any_outdated,
     _rreplace,
     version_up,
-    switch_item,
     _get_host_name,
-    get_asset,
     get_version_from_path,
     get_last_version_from_path,
-    get_subsets,
-    get_linked_assets,
     BuildWorkfile,
     ffprobe_streams,
     source_hash,
-    get_latest_version
 )
 
 __all__ = [
     "get_avalon_database",
     "set_io_database",
+
+    "is_latest",
+    "any_outdated",
+    "switch_item",
+    "get_asset",
+    "get_hierarchy",
+    "get_subsets",
+    "get_linked_assets",
+    "get_latest_version",
 
     "PypeHook",
     "execute_hook",
